+++ Console/Command/AkeneoConnectorImportCommand.php
@@ -69,10 +69,11 @@ class AkeneoConnectorImportCommand extends Command
     protected function configure()
     {
         $this->setName('akeneo_connector:import')->setDescription('Import Akeneo data to Magento')->addOption(
-                self::IMPORT_CODE,
-                null,
-                InputOption::VALUE_REQUIRED
-            );
+            self::IMPORT_CODE,
+            null,
+            InputOption::VALUE_REQUIRED,
+            'Code of import job to run. To run multiple jobs consecutively, use comma-separated import job codes'
+        );
     }
 
     /**